pyproject.toml
[project]
name = "genta-automl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "absl-py==2.2.2",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.18",
    "aiosignal==1.3.2",
    "alembic==1.15.2",
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asttokens==3.0.0",
    "async-timeout==5.0.1",
    "attrs==25.3.0",
    "backcall==0.2.0",
    "bcrypt==4.3.0",
    "beautifulsoup4==4.13.4",
    "bleach==6.2.0",
    "blinker==1.9.0",
    "bokeh==3.7.3",
    "cachetools==5.5.2",
    "catboost==1.2.8",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "cloudpickle==3.1.1",
    "colorama==0.4.6",
    "colorlog==6.9.0",
    "contourpy==1.3.2",
    "cryptography==45.0.2",
    "cycler==0.12.1",
    "daal==2025.5.0",
    "dask==2025.5.1",
    "databricks-sdk==0.53.0",
    "datasets==3.6.0",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "deprecated==1.2.18",
    "dill==0.3.8",
    "distributed==2025.5.1",
    "docker==7.1.0",
    "docopt==0.6.2",
    "eli5==0.16.0",
    "exceptiongroup==1.3.0",
    "executing==2.2.0",
    "fastapi==0.115.12",
    "fastjsonschema==2.21.1",
    "filelock==3.18.0",
    "flaml==2.3.4",
    "flask==3.1.1",
    "fonttools==4.58.0",
    "frozenlist==1.6.0",
    "fsspec==2025.3.0",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "google-auth==2.40.1",
    "graphene==3.4.3",
    "graphql-core==3.2.6",
    "graphql-relay==3.2.0",
    "graphviz==0.20.3",
    "greenlet==3.2.2",
    "h11==0.16.0",
    "hdbscan==0.8.40",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface-hub==0.31.4",
    "idna==3.10",
    "imageio==2.37.0",
    "importlib-metadata==8.6.1",
    "iniconfig==2.1.0",
    "intel-cmplr-lib-ur==2025.1.1",
    "intel-openmp==2025.1.1",
    "ipython==8.12.3",
    "itsdangerous==2.2.0",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "joblib==1.5.0",
    "jsonform==0.0.2",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2025.4.1",
    "jsonsir==0.0.2",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "jupyterlab-pygments==0.3.0",
    "jwt==1.3.1",
    "kiwisolver==1.4.8",
    "lazy-loader==0.4",
    "lightgbm==4.6.0",
    "lime==0.2.0.1",
    "llvmlite==0.44.0",
    "locket==1.0.0",
    "lz4==4.4.4",
    "mako==1.3.10",
    "markdown==3.8",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "matplotlib==3.10.3",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "memory-profiler==0.61.0",
    "mistune==3.1.3",
    "mkl==2025.1.0",
    "mlflow==2.22.0",
    "mlflow-skinny==2.22.0",
    "mpmath==1.3.0",
    "msgpack==1.1.0",
    "multidict==6.4.4",
    "multiprocess==0.70.16",
    "narwhals==1.40.0",
    "nbclient==0.10.2",
    "nbconvert==7.16.6",
    "nbformat==5.10.4",
    "networkx==3.4.2",
    "nltk==3.9.1",
    "numba==0.61.2",
    "numpy==2.2.6",
    "onnx==1.18.0",
    "onnxconverter-common==1.13.0",
    "opentelemetry-api==1.33.1",
    "opentelemetry-sdk==1.33.1",
    "opentelemetry-semantic-conventions==0.54b1",
    "openvino==2025.1.0",
    "openvino-telemetry==2025.1.0",
    "optuna==4.3.0",
    "packaging==24.2",
    "pandas==2.2.3",
    "pandocfilters==1.5.1",
    "parso==0.8.4",
    "partd==1.4.2",
    "pickleshare==0.7.5",
    "pillow==11.2.1",
    "pip==23.0.1",
    "pipreqs==0.5.0",
    "platformdirs==4.3.8",
    "plotly==6.1.1",
    "pluggy==1.6.0",
    "polars==1.29.0",
    "prompt-toolkit==3.0.51",
    "propcache==0.3.1",
    "protobuf==6.31.0",
    "psutil==7.0.0",
    "pure-eval==0.2.3",
    "py-cpuinfo==9.0.0",
    "pyaml==25.1.0",
    "pyarrow==19.0.1",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.4",
    "pydantic-core==2.33.2",
    "pydeck==0.9.1",
    "pygments==2.19.1",
    "pyjwt==2.10.1",
    "pyparsing==3.2.3",
    "pytest==8.3.5",
    "pytest-asyncio==0.26.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-easyconfig==0.1.7",
    "python-multipart==0.0.20",
    "python-resources==0.3",
    "pytz==2025.2",
    "pywin32==310",
    "pyyaml==6.0.2",
    "pyzmq==26.4.0",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "resource==0.2.1",
    "rich==14.0.0",
    "rouge-score==0.1.2",
    "rpds-py==0.25.0",
    "rsa==4.9.1",
    "safetensors==0.5.3",
    "scikit-image==0.25.2",
    "scikit-learn==1.6.1",
    "scikit-learn-intelex==2025.5.0",
    "scikit-optimize==0.10.2",
    "scipy==1.15.3",
    "seaborn==0.13.2",
    "seqeval==1.2.2",
    "setuptools==65.5.0",
    "shap==0.47.2",
    "six==1.17.0",
    "skl2onnx==1.18.0",
    "sklearn2pmml==0.117.0",
    "slicer==0.0.8",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "soupsieve==2.7",
    "sqlalchemy==2.0.41",
    "sqlparse==0.5.3",
    "stack-data==0.6.3",
    "starlette==0.46.2",
    "streamlit==1.45.1",
    "sympy==1.14.0",
    "tabulate==0.9.0",
    "tbb==2022.1.0",
    "tblib==3.1.0",
    "tcmlib==1.3.0",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "tifffile==2025.5.10",
    "tinycss2==1.4.0",
    "tokenizers==0.21.1",
    "toml==0.10.2",
    "tomli==2.2.1",
    "toolz==1.0.0",
    "torch==2.7.0",
    "tornado==6.5",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "transformers==4.52.1",
    "treelite==4.4.1",
    "treelite-runtime==3.9.1",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "tzdata==2025.2",
    "umf==0.10.0",
    "urllib3==2.4.0",
    "uv==0.7.6",
    "uvicorn==0.34.2",
    "waitress==3.0.2",
    "watchdog==6.0.0",
    "wcwidth==0.2.13",
    "webencodings==0.5.1",
    "werkzeug==3.1.3",
    "wrapt==1.17.2",
    "xgboost==3.0.1",
    "xxhash==3.5.0",
    "xyzservices==2025.4.0",
    "yarg==0.1.9",
    "yarl==1.20.0",
    "zict==3.0.0",
    "zipp==3.21.0",
]
